Changed Barometer design to use closures instead of BarometerDelegate protocol
Gave some thought as to whether or not using a delegate protocol was the best way to go.  I came to the following conclusions:

1. Delegates make more sense when they perform a task on behalf of the class (as with the UITableViewDelegate protocol).  This was not the purpose of the BarometerDelegateProtocol.  Rather, it was intended to provide a call back to notify of pressure changes.

2. In order to prevent strong reference cycles, the delegate protocol must be limited to a class protocol (you cannot have a weak reference to a struct).

Therefore I decided to redesign the Barometer protocol to use closures instead of a delegate.

I am happy I did.  In addition to making more sense, it really improved the readability and reduced the amount of code.